fix rest of fast refresh issues
