+++ packages/react/src/primitives/headless.test.tsx
@@ -217,6 +217,7 @@ test('Prompt changes trigger user-defined callbacks', async () => {
 
   await user.type(input, 'test');
   expect(input).toHaveValue('test');
+  // eslint-disable-next-line promise/no-callback-in-promise
   expect(cb).toHaveBeenCalled();
 });
