+++ packages/web/src/index.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable react-refresh/only-export-components */
+
 import {
   Markprompt,
   openMarkprompt,
